Update book structure
I changed `divisions` to `subchapters` and `articles` to `sections`, according to my new understanding of the book structure.
I understood that this might be a better way to structure the book, following the example of StandardEbooks' alexander-berkman/the-bolshevik-myth.
In the ToC, I've also moved the parts structure inside the halftitlepage, following the example of other books I checked.